Move HistoriesService and HistoryContentsService to services package
Also move some of the shared pydantic models from managers package to schema package.
