+++ lib/galaxy/managers/sharable.py
@@ -11,21 +11,14 @@ A sharable Galaxy object:
 """
 import logging
 import re
-from enum import Enum
 from typing import (
     List,
     Optional,
     Set,
     Tuple,
     Type,
-    Union,
 )
 
-from pydantic import (
-    BaseModel,
-    Extra,
-    Field,
-)
 from sqlalchemy import (
     false,
     true,
@@ -45,122 +38,18 @@ from galaxy.model import (
     UserShareAssociation,
 )
 from galaxy.schema.fields import EncodedDatabaseIdField
+from galaxy.schema.schema import (
+    SetSlugPayload,
+    ShareWithExtra,
+    ShareWithPayload,
+    ShareWithStatus,
+    SharingOptions,
+    SharingStatus,
+)
 from galaxy.structured_app import MinimalManagerApp
 from galaxy.util import ready_name_for_url
 
 log = logging.getLogger(__name__)
-
-UserIdentifier = Union[EncodedDatabaseIdField, str]
-
-
-class SharingOptions(str, Enum):
-    """Options for sharing resources that may have restricted access to all or part of their contents."""
-    make_public = "make_public"
-    make_accessible_to_shared = "make_accessible_to_shared"
-    no_changes = "no_changes"
-
-
-class ShareWithExtra(BaseModel):
-    can_share: bool = Field(
-        False,
-        title="Can Share",
-        description="Indicates whether the resource can be directly shared or requires further actions.",
-    )
-
-    class Config:
-        extra = Extra.allow
-
-
-class ShareWithPayload(BaseModel):
-    user_ids: List[UserIdentifier] = Field(
-        ...,
-        title="User Identifiers",
-        description=(
-            "A collection of encoded IDs (or email addresses) of users "
-            "that this resource will be shared with."
-        ),
-    )
-    share_option: Optional[SharingOptions] = Field(
-        None,
-        title="Share Option",
-        description=(
-            "User choice for sharing resources which its contents may be restricted:\n"
-            " - None: The user did not choose anything yet or no option is needed.\n"
-            f" - {SharingOptions.make_public}: The contents of the resource will be made publicly accessible.\n"
-            f" - {SharingOptions.make_accessible_to_shared}: This will automatically create a new `sharing role` allowing protected contents to be accessed only by the desired users.\n"
-            f" - {SharingOptions.no_changes}: This won't change the current permissions for the contents. The user which this resource will be shared may not be able to access all its contents.\n"
-        ),
-    )
-
-
-class SetSlugPayload(BaseModel):
-    new_slug: str = Field(
-        ...,
-        title="New Slug",
-        description="The slug that will be used to access this shared item.",
-    )
-
-
-class UserEmail(BaseModel):
-    id: EncodedDatabaseIdField = Field(
-        ...,
-        title="User ID",
-        description="The encoded ID of the user.",
-    )
-    email: str = Field(
-        ...,
-        title="Email",
-        description="The email of the user.",
-    )
-
-
-class SharingStatus(BaseModel):
-    id: EncodedDatabaseIdField = Field(
-        ...,
-        title="ID",
-        description="The encoded ID of the resource to be shared.",
-    )
-    title: str = Field(
-        ...,
-        title="Title",
-        description="The title or name of the resource.",
-    )
-    importable: bool = Field(
-        ...,
-        title="Importable",
-        description="Whether this resource can be published using a link.",
-    )
-    published: bool = Field(
-        ...,
-        title="Published",
-        description="Whether this resource is currently published.",
-    )
-    users_shared_with: List[UserEmail] = Field(
-        [],
-        title="Users shared with",
-        description="The list of encoded ids for users the resource has been shared.",
-    )
-    username_and_slug: Optional[str] = Field(
-        None,
-        title="Username and slug",
-        description="The relative URL in the form of /u/{username}/{resource_single_char}/{slug}",
-    )
-
-
-class ShareWithStatus(SharingStatus):
-    errors: List[str] = Field(
-        [],
-        title="Errors",
-        description="Collection of messages indicating that the resource was not shared with some (or all users) due to an error.",
-    )
-    extra: Optional[ShareWithExtra] = Field(
-        None,
-        title="Extra",
-        description=(
-            "Optional extra information about this shareable resource that may be of interest. "
-            "The contents of this field depend on the particular resource."
-        ),
-    )
 
 
 class SharableModelManager(base.ModelManager, secured.OwnableManagerMixin, secured.AccessibleManagerMixin,
