+++ lib/galaxy/schema/schema.py
@@ -2364,3 +2364,161 @@ class DeleteHDCAResult(Model):
         title="Deleted",
         description="True if the collection was successfully deleted.",
     )
+
+
+AnyHDA = Union[HDASummary, HDADetailed, HDABeta]
+AnyHistoryContentItem = Union[AnyHDA, HDCASummary, HDCADetailed, HDCABeta]
+
+AnyJobStateSummary = Union[
+    JobStateSummary,
+    ImplicitCollectionJobsStateSummary,
+    WorkflowInvocationStateSummary,
+]
+
+HistoryArchiveExportResult = Union[JobExportHistoryArchiveModel, JobIdResponse]
+
+# Sharing -----------------------------------------------------------------
+
+
+class SharingOptions(str, Enum):
+    """Options for sharing resources that may have restricted access to all or part of their contents."""
+    make_public = "make_public"
+    make_accessible_to_shared = "make_accessible_to_shared"
+    no_changes = "no_changes"
+
+
+class ShareWithExtra(BaseModel):
+    can_share: bool = Field(
+        False,
+        title="Can Share",
+        description="Indicates whether the resource can be directly shared or requires further actions.",
+    )
+
+    class Config:
+        extra = Extra.allow
+
+
+UserIdentifier = Union[EncodedDatabaseIdField, str]
+
+
+class ShareWithPayload(BaseModel):
+    user_ids: List[UserIdentifier] = Field(
+        ...,
+        title="User Identifiers",
+        description=(
+            "A collection of encoded IDs (or email addresses) of users "
+            "that this resource will be shared with."
+        ),
+    )
+    share_option: Optional[SharingOptions] = Field(
+        None,
+        title="Share Option",
+        description=(
+            "User choice for sharing resources which its contents may be restricted:\n"
+            " - None: The user did not choose anything yet or no option is needed.\n"
+            f" - {SharingOptions.make_public}: The contents of the resource will be made publicly accessible.\n"
+            f" - {SharingOptions.make_accessible_to_shared}: This will automatically create a new `sharing role` allowing protected contents to be accessed only by the desired users.\n"
+            f" - {SharingOptions.no_changes}: This won't change the current permissions for the contents. The user which this resource will be shared may not be able to access all its contents.\n"
+        ),
+    )
+
+
+class SetSlugPayload(BaseModel):
+    new_slug: str = Field(
+        ...,
+        title="New Slug",
+        description="The slug that will be used to access this shared item.",
+    )
+
+
+class UserEmail(BaseModel):
+    id: EncodedDatabaseIdField = Field(
+        ...,
+        title="User ID",
+        description="The encoded ID of the user.",
+    )
+    email: str = Field(
+        ...,
+        title="Email",
+        description="The email of the user.",
+    )
+
+
+class SharingStatus(BaseModel):
+    id: EncodedDatabaseIdField = Field(
+        ...,
+        title="ID",
+        description="The encoded ID of the resource to be shared.",
+    )
+    title: str = Field(
+        ...,
+        title="Title",
+        description="The title or name of the resource.",
+    )
+    importable: bool = Field(
+        ...,
+        title="Importable",
+        description="Whether this resource can be published using a link.",
+    )
+    published: bool = Field(
+        ...,
+        title="Published",
+        description="Whether this resource is currently published.",
+    )
+    users_shared_with: List[UserEmail] = Field(
+        [],
+        title="Users shared with",
+        description="The list of encoded ids for users the resource has been shared.",
+    )
+    username_and_slug: Optional[str] = Field(
+        None,
+        title="Username and slug",
+        description="The relative URL in the form of /u/{username}/{resource_single_char}/{slug}",
+    )
+
+
+class ShareWithStatus(SharingStatus):
+    errors: List[str] = Field(
+        [],
+        title="Errors",
+        description="Collection of messages indicating that the resource was not shared with some (or all users) due to an error.",
+    )
+    extra: Optional[ShareWithExtra] = Field(
+        None,
+        title="Extra",
+        description=(
+            "Optional extra information about this shareable resource that may be of interest. "
+            "The contents of this field depend on the particular resource."
+        ),
+    )
+
+
+class HDABasicInfo(BaseModel):
+    id: EncodedDatabaseIdField
+    name: str
+
+
+class ShareHistoryExtra(ShareWithExtra):
+    can_change: List[HDABasicInfo] = Field(
+        [],
+        title="Can Change",
+        description=(
+            "A collection of datasets that are not accessible by one or more of the target users "
+            "and that can be made accessible for others by the user sharing the history."
+        ),
+    )
+    cannot_change: List[HDABasicInfo] = Field(
+        [],
+        title="Cannot Change",
+        description=(
+            "A collection of datasets that are not accessible by one or more of the target users "
+            "and that cannot be made accessible for others by the user sharing the history."
+        ),
+    )
+    accessible_count: int = Field(
+        0,
+        title="Accessible Count",
+        description=(
+            "The number of datasets in the history that are public or accessible by all the target users."
+        ),
+    )
